--- packages/queue/src/runtime/empty-registry.ts
@@ -1,5 +1,0 @@
-import type { QueueDefinitionRegistry } from "../types.ts"
-
-const registry: QueueDefinitionRegistry = {}
-
-export default registry
