+++ packages/queue/test/providers.test.ts
@@ -162,14 +162,11 @@ describe("Vercel provider", () => {
     })
   })
 
-  it("resolves named Vercel clients through getQueue", async () => {
+  it("injects topic from queue name through getQueue", async () => {
     const send = vi.fn(async () => ({ messageId: "msg_vercel" }))
     setQueueRuntimeConfig({
       provider: {
-        client: {
-          handleCallback: vi.fn(),
-          send,
-        },
+        client: { handleCallback: vi.fn(), send },
         provider: "vercel",
       },
     })
